requirements.txt
Django==3.2.23
djongo==1.3.6
dnspython
pymongo[srv]
whitenoise
django-environ
gunicorn
Pillow
cloudinary
django-cloudinary-storage
pytz
asgiref
sqlparse==0.2.4
python-decouple
